fix(html): move metadata handling to own script
